core: extract tree logic to improve code organization
Extract tree representation from run to a dedicated tree package to
clarify responsibilities.
